# zhipin-kratos/api/admin/admin_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.0
// source: admin/admin.proto

package admin

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationAdminServiceGetPositionDetail = "/api.admin.AdminService/GetPositionDetail"
const OperationAdminServiceGetStats = "/api.admin.AdminService/GetStats"
const OperationAdminServiceListAllPositions = "/api.admin.AdminService/ListAllPositions"
const OperationAdminServiceListTodayPositions = "/api.admin.AdminService/ListTodayPositions"

type AdminServiceHTTPServer interface {
	// GetPositionDetail Position detail with match items and communications
	GetPositionDetail(context.Context, *GetPositionDetailReq) (*AdminPositionDetailResp, error)
	// GetStats Stats overview — count by status
	GetStats(context.Context, *GetStatsReq) (*GetStatsResp, error)
	// ListAllPositions All positions with pagination and status filter
	ListAllPositions(context.Context, *ListAllPositionsReq) (*AdminPositionListResp, error)
	// ListTodayPositions Today's positions — most important for the admin to see progress
	ListTodayPositions(context.Context, *ListTodayPositionsReq) (*AdminPositionListResp, error)
}

func RegisterAdminServiceHTTPServer(s *http.Server, srv AdminServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/positions/today", _AdminService_ListTodayPositions0_HTTP_Handler(srv))
	r.GET("/admin/positions", _AdminService_ListAllPositions0_HTTP_Handler(srv))
	r.GET("/admin/position/{id}", _AdminService_GetPositionDetail0_HTTP_Handler(srv))
	r.GET("/admin/stats", _AdminService_GetStats0_HTTP_Handler(srv))
}

func _AdminService_ListTodayPositions0_HTTP_Handler(srv AdminServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListTodayPositionsReq
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminServiceListTodayPositions)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListTodayPositions(ctx, req.(*ListTodayPositionsReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AdminPositionListResp)
		return ctx.Result(200, reply)
	}
}

func _AdminService_ListAllPositions0_HTTP_Handler(srv AdminServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListAllPositionsReq
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminServiceListAllPositions)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListAllPositions(ctx, req.(*ListAllPositionsReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AdminPositionListResp)
		return ctx.Result(200, reply)
	}
}

func _AdminService_GetPositionDetail0_HTTP_Handler(srv AdminServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetPositionDetailReq
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminServiceGetPositionDetail)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetPositionDetail(ctx, req.(*GetPositionDetailReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AdminPositionDetailResp)
		return ctx.Result(200, reply)
	}
}

func _AdminService_GetStats0_HTTP_Handler(srv AdminServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetStatsReq
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminServiceGetStats)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetStats(ctx, req.(*GetStatsReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetStatsResp)
		return ctx.Result(200, reply)
	}
}

type AdminServiceHTTPClient interface {
	// GetPositionDetail Position detail with match items and communications
	GetPositionDetail(ctx context.Context, req *GetPositionDetailReq, opts ...http.CallOption) (rsp *AdminPositionDetailResp, err error)
	// GetStats Stats overview — count by status
	GetStats(ctx context.Context, req *GetStatsReq, opts ...http.CallOption) (rsp *GetStatsResp, err error)
	// ListAllPositions All positions with pagination and status filter
	ListAllPositions(ctx context.Context, req *ListAllPositionsReq, opts ...http.CallOption) (rsp *AdminPositionListResp, err error)
	// ListTodayPositions Today's positions — most important for the admin to see progress
	ListTodayPositions(ctx context.Context, req *ListTodayPositionsReq, opts ...http.CallOption) (rsp *AdminPositionListResp, err error)
}

type AdminServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewAdminServiceHTTPClient(client *http.Client) AdminServiceHTTPClient {
	return &AdminServiceHTTPClientImpl{client}
}

// GetPositionDetail Position detail with match items and communications
func (c *AdminServiceHTTPClientImpl) GetPositionDetail(ctx context.Context, in *GetPositionDetailReq, opts ...http.CallOption) (*AdminPositionDetailResp, error) {
	var out AdminPositionDetailResp
	pattern := "/admin/position/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAdminServiceGetPositionDetail))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetStats Stats overview — count by status
func (c *AdminServiceHTTPClientImpl) GetStats(ctx context.Context, in *GetStatsReq, opts ...http.CallOption) (*GetStatsResp, error) {
	var out GetStatsResp
	pattern := "/admin/stats"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAdminServiceGetStats))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListAllPositions All positions with pagination and status filter
func (c *AdminServiceHTTPClientImpl) ListAllPositions(ctx context.Context, in *ListAllPositionsReq, opts ...http.CallOption) (*AdminPositionListResp, error) {
	var out AdminPositionListResp
	pattern := "/admin/positions"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAdminServiceListAllPositions))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListTodayPositions Today's positions — most important for the admin to see progress
func (c *AdminServiceHTTPClientImpl) ListTodayPositions(ctx context.Context, in *ListTodayPositionsReq, opts ...http.CallOption) (*AdminPositionListResp, error) {
	var out AdminPositionListResp
	pattern := "/admin/positions/today"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAdminServiceListTodayPositions))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}