# zhipin-kratos/api/zhipin/blacklist_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.0
// source: zhipin/blacklist.proto

package zhipin

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationBlacklistServiceAddBlacklist = "/api.zhipin.BlacklistService/AddBlacklist"
const OperationBlacklistServiceCheckBlacklist = "/api.zhipin.BlacklistService/CheckBlacklist"
const OperationBlacklistServiceListBlacklist = "/api.zhipin.BlacklistService/ListBlacklist"
const OperationBlacklistServiceRemoveBlacklist = "/api.zhipin.BlacklistService/RemoveBlacklist"

type BlacklistServiceHTTPServer interface {
	AddBlacklist(context.Context, *AddBlacklistReq) (*BlacklistResp, error)
	CheckBlacklist(context.Context, *CheckBlacklistReq) (*CheckBlacklistResp, error)
	ListBlacklist(context.Context, *ListBlacklistReq) (*ListBlacklistResp, error)
	RemoveBlacklist(context.Context, *RemoveBlacklistReq) (*RemoveBlacklistResp, error)
}

func RegisterBlacklistServiceHTTPServer(s *http.Server, srv BlacklistServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/api/blacklist", _BlacklistService_AddBlacklist0_HTTP_Handler(srv))
	r.GET("/api/blacklist/check/{company}", _BlacklistService_CheckBlacklist0_HTTP_Handler(srv))
	r.GET("/api/blacklist", _BlacklistService_ListBlacklist0_HTTP_Handler(srv))
	r.DELETE("/api/blacklist/{id}", _BlacklistService_RemoveBlacklist0_HTTP_Handler(srv))
}

func _BlacklistService_AddBlacklist0_HTTP_Handler(srv BlacklistServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AddBlacklistReq
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBlacklistServiceAddBlacklist)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.AddBlacklist(ctx, req.(*AddBlacklistReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*BlacklistResp)
		return ctx.Result(200, reply)
	}
}

func _BlacklistService_CheckBlacklist0_HTTP_Handler(srv BlacklistServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CheckBlacklistReq
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBlacklistServiceCheckBlacklist)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CheckBlacklist(ctx, req.(*CheckBlacklistReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CheckBlacklistResp)
		return ctx.Result(200, reply)
	}
}

func _BlacklistService_ListBlacklist0_HTTP_Handler(srv BlacklistServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListBlacklistReq
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBlacklistServiceListBlacklist)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListBlacklist(ctx, req.(*ListBlacklistReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListBlacklistResp)
		return ctx.Result(200, reply)
	}
}

func _BlacklistService_RemoveBlacklist0_HTTP_Handler(srv BlacklistServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RemoveBlacklistReq
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBlacklistServiceRemoveBlacklist)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RemoveBlacklist(ctx, req.(*RemoveBlacklistReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RemoveBlacklistResp)
		return ctx.Result(200, reply)
	}
}

type BlacklistServiceHTTPClient interface {
	AddBlacklist(ctx context.Context, req *AddBlacklistReq, opts ...http.CallOption) (rsp *BlacklistResp, err error)
	CheckBlacklist(ctx context.Context, req *CheckBlacklistReq, opts ...http.CallOption) (rsp *CheckBlacklistResp, err error)
	ListBlacklist(ctx context.Context, req *ListBlacklistReq, opts ...http.CallOption) (rsp *ListBlacklistResp, err error)
	RemoveBlacklist(ctx context.Context, req *RemoveBlacklistReq, opts ...http.CallOption) (rsp *RemoveBlacklistResp, err error)
}

type BlacklistServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewBlacklistServiceHTTPClient(client *http.Client) BlacklistServiceHTTPClient {
	return &BlacklistServiceHTTPClientImpl{client}
}

func (c *BlacklistServiceHTTPClientImpl) AddBlacklist(ctx context.Context, in *AddBlacklistReq, opts ...http.CallOption) (*BlacklistResp, error) {
	var out BlacklistResp
	pattern := "/api/blacklist"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBlacklistServiceAddBlacklist))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BlacklistServiceHTTPClientImpl) CheckBlacklist(ctx context.Context, in *CheckBlacklistReq, opts ...http.CallOption) (*CheckBlacklistResp, error) {
	var out CheckBlacklistResp
	pattern := "/api/blacklist/check/{company}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBlacklistServiceCheckBlacklist))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BlacklistServiceHTTPClientImpl) ListBlacklist(ctx context.Context, in *ListBlacklistReq, opts ...http.CallOption) (*ListBlacklistResp, error) {
	var out ListBlacklistResp
	pattern := "/api/blacklist"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBlacklistServiceListBlacklist))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BlacklistServiceHTTPClientImpl) RemoveBlacklist(ctx context.Context, in *RemoveBlacklistReq, opts ...http.CallOption) (*RemoveBlacklistResp, error) {
	var out RemoveBlacklistResp
	pattern := "/api/blacklist/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBlacklistServiceRemoveBlacklist))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}