# zhipin-kratos/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /admin/position/{id}:
        get:
            tags:
                - AdminService
            description: Position detail with match items and communications
            operationId: AdminService_GetPositionDetail
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.admin.AdminPositionDetailResp'
    /admin/positions:
        get:
            tags:
                - AdminService
            description: All positions with pagination and status filter
            operationId: AdminService_ListAllPositions
            parameters:
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.admin.AdminPositionListResp'
...
    /api/blacklist:
        get:
            tags:
                - BlacklistService
            operationId: BlacklistService_ListBlacklist
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.ListBlacklistResp'
        post:
            tags:
                - BlacklistService
            operationId: BlacklistService_AddBlacklist
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.zhipin.AddBlacklistReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.BlacklistResp'
    /api/blacklist/check/{company}:
        get:
            tags:
                - BlacklistService
            operationId: BlacklistService_CheckBlacklist
            parameters:
                - name: company
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.CheckBlacklistResp'
    /api/blacklist/{id}:
        delete:
            tags:
                - BlacklistService
            operationId: BlacklistService_RemoveBlacklist
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.RemoveBlacklistResp'
...
components:
    schemas:
        api.admin.AdminCommItem:
            type: object
            properties:
                eventType:
                    type: integer
                    format: int32
                eventTime:
                    type: string
                    format: date-time
                content:
                    type: string
                direction:
                    type: integer
                    format: int32
        api.admin.AdminMatchItem:
            type: object
            properties:
                requirement:
                    type: string
                matchStatus:
                    type: integer
                    format: int32
                resumePoint:
                    type: string
                remark:
                    type: string
        api.admin.AdminPositionDetailResp:
            type: object
            properties:
                position:
                    $ref: '#/components/schemas/api.admin.AdminPositionItem'
                matchItems:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.admin.AdminMatchItem'
                communications:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.admin.AdminCommItem'
                duties:
                    type: string
                requirements:
                    type: string
                notes:
                    type: string
                skipReason:
                    type: string
        api.admin.AdminPositionItem:
            type: object
            properties:
                id:
                    type: string
                jobId:
                    type: string
                title:
                    type: string
                company:
                    type: string
                salaryRange:
                    type: string
                salaryMin:
                    type: integer
                    format: int32
                salaryMax:
                    type: integer
                    format: int32
                city:
                    type: string
                recruiter:
                    type: string
                isHunter:
                    type: boolean
                status:
                    type: integer
                    format: int32
                matchRate:
                    type: integer
                    format: int32
                lastResume:
                    type: string
                lastCommAt:
                    type: string
                lastCommDir:
                    type: integer
                    format: int32
                link:
                    type: string
        api.admin.AdminPositionListResp:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.admin.AdminPositionItem'
                total:
                    type: integer
                    format: int32
...
        api.zhipin.AddBlacklistReq:
            type: object
            properties:
                company:
                    type: string
                reason:
                    type: string
...
        api.zhipin.BlacklistResp:
            type: object
            properties:
                id:
                    type: string
                company:
                    type: string
                reason:
                    type: string
        api.zhipin.CheckBlacklistResp:
            type: object
            properties:
                blacklisted:
                    type: boolean
...
        api.zhipin.ListBlacklistResp:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.zhipin.BlacklistResp'
...
        api.zhipin.RemoveBlacklistResp:
            type: object
            properties: {}
...